Graceful sidecar support
TL;DR Tekton Pipelines now gracefully start and stop with sidecars
running alongside TaskRun pods.

Long Version :-

TaskRun pods can have sidecars injected by Admission Controllers. This
is common practice with Istio where a proxy sidecar is injected into pods
so that they can be included as part of a service mesh. Unfortunately
there is no built-in Kubernetes mechanism to set the lifetime of a
sidecar container to match that of the pod's "primary" container. In
practice this means that pods injected with sidecars will not stop until
both the primary and all sidecar containers are stopped.

Prior to this commit injected sidecars would cause a TaskRun pod to run
forever, even when all Step containers of a Task had completed. This
commit introduces mechanisms to start and stop sidecars gracefully
meaning that a) Step containers wait until all sidecars are in a ready
state before starting and b) Sidecar containers are stopped when a
TaskRun's Step containers are done.

Co-authored-by: Jose Blas Camacho Taboada <joseblas@gmail.com>
Co-authored-by: Scott Seaward <sbws@google.com>