Add http json log exporter to entrypoint
This change introduces a log exporter to the entrypoint that POSTs log
lines from the entrypoint's child process to a declared logging
endpoint. The intention is for this logging endpoint to be a sidecar
running in the same Pod as the entrypoint binary.

The logging entrypoint is configured with the --log_url flag. Additional
flags can be provided to specify the --pipeline, --pipelinerun, --task
and --taskrun that this entrypoint is emitting log lines for. These
string flags will end up in the JSON objects being sent out by the log
exporter.

WIP.

- In manual testing I've noticed the entrypoint binary pegs 2 full CPUs
while exporting the log lines. I want to get to the bottom of why this
is.