Add explicit named anchors to workspace doc
Currently whenever someone updates a heading anywhere in one of our docs
they break any inbound links to that heading from either other docs or
from external websites.

This commit introduces explicit named identifiers for the headings in our
workspaces doc. The names match precisely what we have right now to
support any existing external links. In future if this doc is updated
or the headings are re-worded we can maintain these named anchors
and people will be correctly guided to the right part of the doc.