Remove creds-init and as many mentions of it as possible
This commit removes the creds-init helper binary and all references
to it in our codebase. The pkg/pod/creds_init.go file is updated
to supply credential configuration information that is then passed
to the entrypoint. The entrypoint is now responsible for moving
credentials into the correct locations. The complete order of operations
wrt credentials is the following:

1. Each Step receives a unique volume mount at /tekton/creds
  - This is done so that the directory is world-writeable regardless of
    Step's UID. It is also done so that /tekton/creds is _always_ empty
    when a Step starts.
2. $(credentials.path) is always set to /tekton/creds
3. Secrets annotated as described by our auth.md doc are now mounted
   in to every single Step container.
4. Arguments that used to be passed to creds-init are now passed
   to the entrypoint.
5. The entrypoint copies the credentials out of the secret volumes
   and into /tekton/creds.
  - This is done because I'm assuming we need to honor the
    $(credentials.path) variable as part of our API compatibility policy.
    If this isn't the case then we can skip this part.
6. Finally, entrypoint copies credentials out of /tekton/creds and into
   /tekton/home or $HOME or wherever the container/UID dictates HOME is.

The results here are the same as with creds-init: credentials are mounted
from secrets according to annotations, they're copied into a single location
that is referrable with $(credentials.path), and they're copied into $HOME
so that they're usable by ssh / docker. The primary improvements we've made
here are:

- The credentials are now owned by the UID of the Step.
  * Why? Previously creds-init would write credentials with a single UID, preventing
    Steps from having access to those creds if they ran as a different UID.
- The credentials will now be placed into each Step's $HOME.
  * Why? Previously creds-init assumed that /tekton/home was the HOME directory of
    all Steps. Our intention is to remove this piece of our runtime contract.
    It's not possible for creds-init to ascertain the HOME directory of each
    Step in a Task and so it's unable to successfully copy the creds into all
    the locations that they could be potentially needed.

On first read these changes can appear to reduce the security of sensitive
credentials mounted from Secrets. However, this doesn't seem to be the case :-
in the "old way" creds-init copied credentials out of secret volume mounts and into
/tekton/home, a world-readable shared directory that every Step had
access to. Therefore credentials were by no means "safe". This is still true with
the "new way" as well - all credentials are mounted as world-readable
and they're copied into directories used by all the Steps of a Task. This commit
is not intended to address the glaring security problems with our creds-init
process but instead to make sure that, once we remove the /tekton/home default
HOME dir, that we do as much as possible to maintain the creds-init contract.

One response to this commit might be, "well, why do we even want this process?"
and that's a totally fair question to ask. If we decide to simply drop the
creds-init flow when we switch the HOME default then a lot of things suddenly
get much easier. The question then becomes, "what are we going to offer users
in the way of credentials support in a post-creds-init world?"