Modify pipelinerun reconciler to let CustomTasks return Task YAML
This commit provides a proof-of-concept for "remote task resolution". The user can
specify a task from git, gcs, a tekton bundle registry or a tekton catalog and the
reconciler will execute that task as if it were stored "locally" in the cluster.

This proof-of-concept is a total hack on top of the existing custom tasks
implementation in tandem with the `TaskRef.Bundle` field introduced as part of
Tekton Bundles.

The way this works is that any PipelineTask which refers to a Custom Task apiVersion
and includes a `Bundle` will follow a different protocol than regular Custom Tasks:

1. The PipelineRun reconciler will create the Run as normal.
2. The Custom Task's controller will see the Run and fetch the task YAML from the
remote source.
3. The Custom Task's controller updates the Run with the YAML embedded in the
`status.extraFields` JSON blob.
4. The Custom Task's controller sets the `TaskResolved` Condition to "True".
5. The PipelineRun reconciler sees the updated condition and gets the Task YAML
out of the `extraFields` blob.
6. The PipelineRun reconciler creates a TaskRun with the fetched Task embedded
in the `taskSpec`. Params and workspaces are passed through from the original
PipelineTask.

Included with this commit are the reconciler changes and an example pipeline that
combines Tasks from Tekton Bundles, Git, GCS, and two different catalogs. The Custom
Task Controllers for each of these "sources" will be pushed separately in a commit
to the Experimental repo.