Remove task spec resolution from taskrun reconciler
Prior to this commit the taskrun reconciler was responsible
for resolving task specs from taskRef / taskSpec / bundles.

After this commit the taskrun reconciler ignores any taskrun
that doesn't have the task spec cached in the taskrun's
status. For now that means all taskruns are ignored -
the logic for populating the taskrun status from a task spec
has been removed entirely from the controller.